requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pdfplumber>=0.10.0
fpdf2>=2.7.0
joblib>=1.3.0
regex>=2023.8.0
